new configuration to keep original when converting

+++ src/converter.ts
@@ -1,7 +1,8 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 
-import { ConfigId, getConfig, getJsonFromYaml, getYamlFromJson, showError } from './helpers';
+import { getJsonFromYaml, getYamlFromJson, showError } from './helpers';
+import { ConfigId, Configs, getConfig } from './config';
 
 type ConvertedFile = {
 	oldFileUri: vscode.Uri;
@@ -21,12 +22,13 @@ export class FileConverter {
 	}
 
 	public async convertFiles(files: vscode.Uri[]): Promise<void> {
-		const convertFilePromises = files.map(this.transformAndConvertFile);
+		const shouldKeepOriginalFiles = await this.shouldKeepOriginalFiles(files.length);
+		const convertFilePromises = files.map((file) => this.transformAndConvertFile(shouldKeepOriginalFiles, file));
 		const convertedFiles = await Promise.all(convertFilePromises);
-		await this.showReverterTooltip(convertedFiles);
+		!shouldKeepOriginalFiles && await this.showReverterTooltip(convertedFiles);
 	}
 
-	private transformAndConvertFile = async (oldFileUri: vscode.Uri): Promise<ConvertedFile> => {
+	private transformAndConvertFile = async (shouldRenameFile: boolean, oldFileUri: vscode.Uri): Promise<ConvertedFile> => {
 		const oldFileContent = await vscode.workspace.fs.readFile(oldFileUri);
 		const oldFileExtension = path.extname(oldFileUri.fsPath);
 
@@ -35,7 +37,7 @@ export class FileConverter {
 		const newFileUri = vscode.Uri.file(newFilePath);
 		const newFileContent = FileConverter.getFileConverter(this.convertFromType)(oldFileContent.toString());
 
-		await this.convertFile(oldFileUri, newFileUri, newFileContent);
+		await this.convertFile(shouldRenameFile, oldFileUri, newFileUri, newFileContent);
 
 		return { oldFileUri, oldFileContent, newFileUri };
 	};
@@ -54,7 +56,9 @@ export class FileConverter {
 			return;
 		}
 
-		await Promise.all(convertedFiles.map(this.revertTransformedAndConvertedFile));
+		const shouldKeepOriginalFiles = false; // always revert created file(s)
+		const promises = convertedFiles.map(async (convertedFile) => this.revertTransformedAndConvertedFile(shouldKeepOriginalFiles, convertedFile));
+		await Promise.all(promises);
 
 		const revertedMessage = didConvertSingleFile
 			? 'Successfully reverted converted file'
@@ -63,24 +67,81 @@ export class FileConverter {
 		vscode.window.showInformationMessage(revertedMessage);
 	}
 
-	private revertTransformedAndConvertedFile = async (convertedFile: ConvertedFile) => {
+	private revertTransformedAndConvertedFile = async (shouldKeepOriginalFiles: boolean, convertedFile: ConvertedFile) => {
 		const {
 			oldFileUri: newFileUri,
 			oldFileContent: newFileContent,
 			newFileUri: oldFileUri
 		} = convertedFile;
 
-		await this.convertFile(oldFileUri, newFileUri, newFileContent.toString());
+		await this.convertFile(shouldKeepOriginalFiles, oldFileUri, newFileUri, newFileContent.toString());
 	};
 
-	private convertFile = async (oldFileUri: vscode.Uri, newFileUri: vscode.Uri, newFileContent: string) => {
+	/**
+	 * @param oldFileUri 
+	 * @param shouldRename 
+	 * @param newFileUri 
+	 * @param newFileContent 
+	 */
+	private convertFile = async (shouldKeepOriginalFile: boolean, oldFileUri: vscode.Uri, newFileUri: vscode.Uri, newFileContent: string) => {
+		const newFile = Buffer.from(newFileContent);
+
+		const existentFile = await this.doFileExist(newFileUri);
+		if (existentFile) {
+			vscode.window.showInformationMessage(`file already exist: ${newFileUri}`);
+			return;
+		}
+
+		if (shouldKeepOriginalFile) {
+			try {
+				await vscode.workspace.fs.writeFile(newFileUri, newFile);
+			} catch (error: any) {
+				console.error(error);
+				showError(error);
+			}
+			return;
+		}
+
 		try {
-			await vscode.workspace.fs.writeFile(oldFileUri, Buffer.from(newFileContent));
+			await vscode.workspace.fs.writeFile(oldFileUri, newFile);
 			await vscode.workspace.fs.rename(oldFileUri, newFileUri);
-		} catch (error) {
+		} catch (error: any) {
+			console.error(error);
 			showError(error);
 		}
 	};
+
+	private async doFileExist(fileUri: vscode.Uri): Promise<boolean> {
+		try {
+			await vscode.workspace.fs.readFile(fileUri);
+			return true;
+		} catch (error) {
+			if (error instanceof vscode.FileSystemError) {
+				return false;
+			}
+
+			throw error;
+		}
+	}
+
+	private async shouldKeepOriginalFiles(length: number): Promise<boolean> {
+		const keepOriginalFiles = getConfig<Configs['keepOriginalFiles']>(ConfigId.KeepOriginalFiles);
+
+		if (keepOriginalFiles === 'always') {
+			return true;
+		}
+
+		if (keepOriginalFiles === 'ask') {
+			const message = length === 1
+				? 'Do you want to keep original file?'
+				: 'Do you want to keep original files?';
+			const selection = await vscode.window.showInformationMessage(message, 'Keep', 'Dont keep');
+
+			return selection === 'Keep';
+		}
+
+		return false;
+	}
 
 	private static getFileConverter(convertFromType: ConvertFromType) {
 		return {
